+++ tests/testthat/test-portfolio.R
@@ -96,7 +96,12 @@ test_that(".check_portfolio works as expected", {
     "CRS of x is not EPSG:4326. Attempting to transform."
   )
   mapme_options(verbose = TRUE)
-  expect_message(.check_portfolio(x), "Found a column named 'assetid'.")
+  expect_silent(.check_portfolio(x))
+  x$assetid <- NULL
+  expect_silent(x <- .check_portfolio(x))
+  expect_true("assetid" %in% names(x))
+  x <- do.call(rbind, lapply(1:2, function(y) x))
+  expect_message(.check_portfolio(x), "Found a column named 'assetid' with non-unique identifiers")
 })
 
 test_that("portfolio helpers work as expected", {